Add files via upload
create a personal portfolio website design step by step using HTML CSS and JavaScript. we can use this website to make your online resume or CV. In this website design we will add name, designation of the user. Then we will display the image, skills, work experience and education details. We will also add the list of services and work portfolio on this website. In the footer of this website we will add contact info and contact form